+++ packages/design-system/.eslintrc.json
@@ -1,0 +1,6 @@
+{
+  "rules": {
+    "no-console": "off"
+  },
+  "ignorePatterns": ["dist/**", "dist-types/**"]
+}
