+++ packages/design-system/README.md
@@ -1,0 +1,4 @@
+# Common
+
+## Description
+A package for common utilities that are being used in multiple packages
