+++ packages/design-system/package.json
@@ -1,0 +1,31 @@
+{
+  "name": "@cookie-analysis-tool/design-system",
+  "version": "0.2.0",
+  "description": "A package for common presentational components that are being used in multiple packages",
+  "main": "src/index.js",
+  "source":"./src/index.ts",
+  "customExports": {
+    ".": {
+      "default": "./src/index.ts"
+    }
+  },
+  "scripts": {
+    "build": "tsc",
+    "dev": "tsc-watch"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/GoogleChromeLabs/ps-analysis-tool"
+  },
+  "author": "",
+  "license": "Apache-2.0",
+  "bugs": {
+    "url": "https://github.com/GoogleChromeLabs/ps-analysis-tool/issues"
+  },
+  "homepage": "https://github.com/GoogleChromeLabs/ps-analysis-tool",
+  "dependencies": {
+    "tldts": "^6.0.8"
+  },
+  "devDependencies": {
+  }
+}
