+++ packages/extension/package.json
@@ -27,6 +27,7 @@
     "use-context-selector": "^1.4.1",
     "use-debounce": "^9.0.4",
     "victory": "^36.6.11",
-    "@cookie-analysis-tool/common":"*"
+    "@cookie-analysis-tool/common":"*",
+    "@cookie-analysis-tool/design-system":"*"
   }
 }
