+++ packages/extension/src/view/popup/app.tsx
@@ -25,10 +25,11 @@ import React from 'react';
 import './app.css';
 import { Legend } from './components';
 import { useCookieStore } from './stateProviders/syncCookieStore';
-import { Button, CirclePieChart } from '../design-system/components';
+import { CirclePieChart } from '../design-system/components';
 import { prepareCookieStatsComponents } from '../../utils/prepareCookieStatsComponents';
 import ProgressBar from '../design-system/components/progressBar';
 import { ALLOWED_NUMBER_OF_TABS } from '../../constants';
+import { Button } from '@cookie-analysis-tool/design-system';
 
 const App: React.FC = () => {
   const {
