+++ packages/extension/tsconfig.json
@@ -7,7 +7,8 @@
   "include": [
     "src",
     "../../node_modules/@types/jest/index.d.ts",
-    "src/**/*.json"
+    "src/**/*.json",
+    "src/view/design-system/components/circle"
   ],
   "references": [{ "path": "../common" }]
 }
